✨[feat] LaunchVC 구현
메인VC에 있던 런치 로직 이동
