+++ TripLog/TripLog/MainViewController.swift
@@ -17,36 +17,23 @@ class MainViewController: UIViewController {
     
     private let disposeBag = DisposeBag()
     
-    private let mainVC = CustomTabBarController()
+    private let mainVC: CustomTabBarController
     
-    // MARK: - UI Compnents
-    
-    private lazy var lottieAnimationView = LottieAnimationView(name: "triplog").then {
-        $0.alpha = 1
-        $0.loopMode = .repeat(3) // 애니메이션 재생 횟수
+    // MARK: - Initializer
+    init(coordinator: MainCoordinator) {
+        self.mainVC = .init(coordinator: coordinator)
+        super.init(nibName: nil, bundle: nil)
     }
     
-    private lazy var launchImageView = UIImageView().then {
-        $0.contentMode = .scaleAspectFill
-        $0.backgroundColor = .clear
-        $0.image = UIImage(named: "launchImage")
-        $0.alpha = 1
-    }
-    
-    private let launchTitle = UILabel().then {
-        $0.text = "TripLog"
-        $0.textColor = .white
-        $0.font = .SCDream(size: .title, weight: .bold)
-        $0.textAlignment = .center
-        $0.backgroundColor = .clear
-        $0.alpha = 1
+    @available(*, unavailable)
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
     }
     
     // MARK: - MainViewController LifeCycle
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        
         setupUI()
         
         Task {
@@ -74,39 +61,15 @@ class MainViewController: UIViewController {
 private extension MainViewController {
     
     func setupUI() {
-        // TabBarController 삽입
         addChild(mainVC)
         view.addSubview(mainVC.view)
         mainVC.didMove(toParent: self)
-        
         configureSelf()
-        setupLayout()
-        playLottie()
+        showOnboardingView()
     }
     
     func configureSelf() {
         navigationItem.title = ""
-        [launchImageView,
-         lottieAnimationView,
-         launchTitle
-        ].forEach { view.addSubview($0) }
-    }
-    
-    func setupLayout() {
-        launchImageView.snp.makeConstraints {
-            $0.edges.equalToSuperview()
-        }
-        
-        lottieAnimationView.snp.makeConstraints {
-            $0.horizontalEdges.equalToSuperview().inset(24)
-            $0.height.equalTo(self.view.bounds.height)
-            $0.bottom.equalToSuperview().inset(50)
-        }
-        
-        launchTitle.snp.makeConstraints {
-            $0.bottom.equalTo(self.view.safeAreaLayoutGuide).inset(50)
-            $0.centerX.equalToSuperview()
-        }
     }
     
     /// 메인뷰에 온보딩 뷰를 설정하는 메소드
@@ -139,34 +102,6 @@ private extension MainViewController {
         guard isFirstLaunch else { return }
         UIView.transition(with: view, duration: 0.3, options: .transitionCrossDissolve) {
             self.setupOnboardingView()
-        }
-    }
-    
-    /// Lottie 애니메이션 실행 메소드
-    func playLottie() {
-        lottieAnimationView.play { [weak self] _ in
-            guard let self else { return }
-            
-            // lottie 애니메이션 재생 완료 후 동작
-            UIView.animate(withDuration: 0.3, animations: {
-                [self.lottieAnimationView,
-                 self.launchImageView,
-                 self.launchTitle
-                ].forEach { $0.alpha = 0 }
-                
-                self.showOnboardingView()
-                
-                // view 애니메이션 종료 후 동작
-            }, completion: { _ in
-                [self.lottieAnimationView,
-                 self.launchImageView,
-                 self.launchTitle
-                ].forEach {
-                    $0.isHidden = true
-                    $0.removeFromSuperview()
-                    $0.snp.removeConstraints()
-                }
-            })
         }
     }
 }
